# docs/features/forgot-password/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Forgot / Reset / Change Password API
  version: 0.1.0
  description: |
    Self-service password recovery for job-seekers with a Local (email+password) account, plus an
    authenticated change-password flow. A job-seeker with a Google OAuth account (no
    `passwordHash`) never hits a dead end: both the reset-request and change-password paths
    explain that the account signs in with Google instead of failing silently or with a generic
    error (idea-brief.md §2, §7 Approach C; PRD.md AC-05).

    Security posture (PRD.md §6, §6.1): reset tokens are single-use and expire after 15 minutes;
    reset requests are rate-limited to 3/hour per email; no response ever reveals whether an email
    is registered (AC-02) — except the deliberate, PRD-mandated Google-account exception (AC-05).
    A successful password change (via reset or change-password) invalidates every other active
    session via `User.tokenVersion` (AC-06, ADR-0002).

servers:
  - url: http://localhost:3000
    description: Local (matches this project's existing /api/auth/* mount, no URL version prefix)

paths:
  /api/auth/password-reset/request:
    post:
      summary: Request a password-reset email
      operationId: requestPasswordReset
      description: |
        Always responds 200 to avoid leaking whether an email is registered (AC-02), with one
        deliberate exception: a Google-only account gets an explicit `hint` instead of the generic
        confirmation (AC-05) — never a dead end, per idea-brief.md §7 Approach C.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPasswordResetBody"
            example:
              email: jobseeker@example.test
      responses:
        "200":
          description: >-
            Generic confirmation (registered or not — AC-02) or a Google-account hint (AC-05).
            Never distinguishes "unknown email" from "known email" by status or shape.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestPasswordResetResponse"
              examples:
                genericConfirmation:
                  summary: Known Local account or unknown email (AC-01, AC-02)
                  value:
                    message: If that email is registered, a reset link has been sent.
                    attemptsRemaining: 2
                googleAccountHint:
                  summary: Known Google-only account (AC-05)
                  value:
                    message: This account signs in with Google. Use "Sign in with Google" instead.
                    hint: google_account
        "429":
          description: Reset-request rate limit exceeded for this email (PRD.md §6 NFR — 3/hour)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: password_reset.rate_limited
                message: Too many reset requests for this email. Try again later.
        "4XX":
          $ref: "#/components/responses/Error"

  /api/auth/password-reset/confirm:
    post:
      summary: Set a new password using a reset token
      operationId: confirmPasswordReset
      description: |
        Consumes the single-use reset token atomically (data-model.md — `findOneAndDelete` on
        `PasswordReset`) and, on success, bumps `User.tokenVersion` to invalidate every other
        active session (AC-06).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPasswordResetBody"
            example:
              token: 3f2a9c1e7b4d4a8f9c0e5d6b7a8c9d0e1f4b8a7c6d5e9f0a1b2c3d4e5f6a7b8c
              newPassword: correct-horse-battery-staple
      responses:
        "200":
          description: Password set; every other active session is now invalid (AC-01, AC-06)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmPasswordResetResponse"
              example:
                message: Password updated. Please log in again.
        "400":
          description: Reset link is expired or was already used (AC-03)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: password_reset.invalid_or_expired_token
                message: This reset link is invalid or has expired. Request a new one.
        "4XX":
          $ref: "#/components/responses/Error"

  /api/auth/change-password:
    post:
      summary: Change password while logged in
      operationId: changePassword
      security:
        - CookieAuth: []
      description: |
        Requires the existing `requireAuth` session (`token` cookie). On success bumps
        `User.tokenVersion`, invalidating every other active session (AC-06) — the same
        invalidation the reset-confirm path triggers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
            example:
              currentPassword: old-password-123
              newPassword: correct-horse-battery-staple
      responses:
        "200":
          description: Password changed; every other active session is now invalid (AC-06)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangePasswordResponse"
              example:
                message: Password updated.
        "400":
          description: Current password is incorrect (AC-04) — existing password left unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: auth.invalid_current_password
                message: Your current password is incorrect.
        "409":
          description: Account has no password to change — Google-only account (AC-05)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: auth.google_account_no_password
                message: This account signs in with Google and has no password to change.
        "4XX":
          $ref: "#/components/responses/Error"

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: >-
        httpOnly JWT cookie set by issueSession (src/controllers/auth.controller.ts) — matches
        this project's existing requireAuth middleware; not the skill-default BearerAuth (see
        api-sync-report.md Deviations).

  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: "password_reset.invalid_or_expired_token" }
        message: { type: string }
        details: { type: object }

    RequestPasswordResetBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: "origin: data-model.md User.email (existing, required, unique)"

    RequestPasswordResetResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
        hint:
          type: [string, "null"]
          enum: [google_account, null]
          description: >-
            Set only for a known Google-only account (AC-05) — the sole intentional exception to
            AC-02's existence-hiding rule, called for explicitly by the PRD.
        attemptsRemaining:
          type: integer
          minimum: 0
          description: >-
            origin: data-model.md PasswordReset.attemptsRemaining (reconciled 2026-09-09) — reset
            requests remaining for this email in the current rolling hour window (PRD §6 NFR:
            ≤3/hour). Present on the generic confirmation for registered and unknown emails alike
            (counted in `ResetRequestAttempt`, not `PasswordReset`) — omitting it for unknown emails
            would reveal that no account exists. Absent only on the Google-account hint.

    ConfirmPasswordResetBody:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 64
          maxLength: 64
          pattern: "^[0-9a-f]{64}$"
          description: >-
            origin: data-model.md PasswordReset.tokenHash ("SHA-256 hex digest of the raw token
            emailed to the job-seeker") — this field carries the raw pre-hash token; the model
            never exposes tokenHash itself. Fixed at 64 lowercase hex chars, matching sad.md §6
            US-01's "Generate raw token, hash it (sha256)" step under the common
            crypto.randomBytes(32).toString('hex') convention. confidence: medium (raw-token
            shape confirmed by sequence, but the exact generation call is an implementation
            detail, not itself in data-model.md).
        newPassword:
          type: string
          minLength: 8
          description: >-
            origin: PRD.md §1 Context (8-character minimum, matching register's existing
            PASSWORD_MIN_LENGTH in src/controllers/auth.controller.ts). Write-only — never stored
            raw, only as User.passwordHash.

    ConfirmPasswordResetResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    ChangePasswordBody:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          description: "origin: PRD.md AC-04 — compared against User.passwordHash, never stored."
        newPassword:
          type: string
          minLength: 8
          description: >-
            origin: PRD.md §1 Context (8-character minimum, matching register's existing
            PASSWORD_MIN_LENGTH). Write-only — never stored raw, only as User.passwordHash.

    ChangePasswordResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }